Text/HuggingFace/requirements.txt
azure-functions
requests
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
transformers